pyproject.toml
[project]
name = "ai3-fullstack"
version = "0.1.0"
description = "AI-3: Full Stack AI Engineering"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "voyageai>=0.3.0",
    "pydantic>=2.0.0",
    "chromadb>=0.5.0",
    "numpy>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "arize-phoenix>=13.0.0",
    "arize-phoenix-otel>=0.6.0",
    "openinference-instrumentation-anthropic>=0.1.0",
    "openinference-instrumentation>=0.1.0",
    "nest-asyncio>=1.6.0",
    "streamlit>=1.40.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
cloud = ["pysqlite3-binary>=0.5.0"]
